+++ src/app/constants.ts
@@ -44,6 +44,11 @@ export const pybricksUsbDfuWindowsDriverInstallUrl =
 export const pybricksUsbLinuxUdevRulesUrl =
     'https://github.com/pybricks/support/discussions/688#discussioncomment-3239099';
 
+export const pybricksBleFirmwareRestoreVideoUrl =
+    'https://pybricks.com/install/technic-boost-city/#restoring-the-original-firmware';
+
+export const pybricksDfuRestoreUrl = 'https://dfu.pybricks.com';
+
 /** Pybricks copyright statement. */
 export const pybricksCopyright = 'Copyright (c) 2020-2022 The Pybricks Authors';