+++ src/firmware/restoreOfficialDialog/reducers.ts
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2022 The Pybricks Authors
+
+import reducer from './redux';
+
+export default reducer;